add: remove existing stuff when reading tree (read-tree) to avoid conflicts
This is done so that we won't have any older files left around after a read-tree called.
Before this change, if we save a tree A which contains only a.txt, then we save tree B which contains a.txt and b.txt and then we read-tree A, we will have b.txt left over in working directory.